+++ packages/ethers-ethereum/src/index.test.ts
@@ -43,4 +43,43 @@ describe.each(data)("ethers.js Ethereum", (eth: Ethereum) => {
 		const networkId = await eth.getChainId()
 		expect(networkId).toBe(17)
 	})
+
+	test(`${eth.constructor.name} encode/decode`, async () => {
+		const type = {
+			components: [
+				{
+					name: "payouts",
+					type: "uint256",
+				},
+				{
+					name: "originFeeFirst",
+					type: "uint256",
+				},
+				{
+					name: "marketplaceMarker",
+					type: "bytes",
+				},
+				{
+					name: "r",
+					type: "uint8",
+				},
+			],
+			name: "data",
+			type: "tuple",
+		}
+
+		const data = {
+			payouts: 0x4058,
+			originFeeFirst: 0x1011,
+			marketplaceMarker: "0xabcdef",
+			r: 78,
+		}
+
+		const encoded = eth.encodeParameter(type,	data)
+		const decoded = eth.decodeParameter(type,	encoded)
+		for (const field in data) {
+			//@ts-ignore
+			expect(decoded[field]).toEqual(data[field].toString())
+		}
+	})
 })