+++ packages/sdk/src/config/mainnet.ts
@@ -10,8 +10,9 @@ export const mainnetConfig: EthereumConfig = {
 		v1: toAddress("0x09EaB21c40743B2364b94345419138eF80f39e30"),
 		v2: toAddress("0x9757F2d2b135150BBeb65308D4a91804107cd8D6"),
 		openseaV1: toAddress("0x7be8076f4ea4a4ad08075c2508e481d6c946d12b"),
-		wrapper: toAddress("0x2a7251d1e7d708c507b1b0d3ff328007beecce5d"),
+		wrapper: toAddress("0x2C45Af926d5f62C5935278106800a03eB565778E"),
 		looksrare: toAddress("0x59728544B08AB483533076417FbBB2fD0B17CE3a"),
+		x2y2: toAddress("0x6d7812d41a08bc2a910b562d8b56411964a4ed88"),
 	},
 	transferProxies: {
 		nft: toAddress("0x4fee7b061c97c9c496b01dbce9cdb10c02f0a0be"),
@@ -34,6 +35,10 @@ export const mainnetConfig: EthereumConfig = {
 	cryptoPunks: {
 		marketContract: toAddress("0x6Ba6f2207e343923BA692e5Cae646Fb0F566DB8D"),
 		wrapperContract: toAddress("0x282BDD42f4eb70e7A9D9F40c8fEA0825B7f68C5D"),
+	},
+	sudoswap: {
+		pairFactory: toAddress("0xb16c1342E617A5B6E4b631EB114483FDB289c0A4"),
+		pairRouter: toAddress("0x2b2e8cda09bba9660dca5cb6233787738ad68329"),
 	},
 	weth: toAddress("0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"),
 	auction: ZERO_ADDRESS,