+++ packages/sdk/src/config/polygon.ts
@@ -11,6 +11,7 @@ export const polygonConfig: EthereumConfig = {
 		v2: toAddress("0x12b3897a36fDB436ddE2788C06Eff0ffD997066e"),
 		openseaV1: ZERO_ADDRESS,
 		wrapper: ZERO_ADDRESS,
+		x2y2: ZERO_ADDRESS,
 	},
 	transferProxies: {
 		nft: toAddress("0xd47e14DD9b98411754f722B4c4074e14752Ada7C"),
@@ -32,6 +33,10 @@ export const polygonConfig: EthereumConfig = {
 	cryptoPunks: {
 		marketContract: ZERO_ADDRESS,
 		wrapperContract: ZERO_ADDRESS,
+	},
+	sudoswap: {
+		pairFactory: ZERO_ADDRESS,
+		pairRouter: ZERO_ADDRESS,
 	},
 	weth: toAddress("0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619"),
 	auction: ZERO_ADDRESS,