+++ packages/sdk/src/config/type.ts
@@ -7,6 +7,7 @@ export type ExchangeAddresses = {
 	openseaV1: Address
 	wrapper: Address
 	looksrare?: Address
+	x2y2: Address
 }
 
 export type TransferProxies = {
@@ -34,6 +35,11 @@ export type CryptoPunksConfig = {
 	wrapperContract: Address
 }
 
+export type SudoswapConfig = {
+	pairFactory: Address
+	pairRouter: Address
+}
+
 export type EthereumConfig = {
 	basePath: string
 	chainId: number
@@ -43,6 +49,7 @@ export type EthereumConfig = {
 	openSea: OpenSeaConfig
 	factories: FactoriesAddresses
 	weth: Address
-	auction: Address,
+	auction: Address
 	cryptoPunks: CryptoPunksConfig
+	sudoswap: SudoswapConfig
 }