+++ packages/sdk/src/order/cancel.test.ts
@@ -14,6 +14,7 @@ import { createEthereumApis } from "../common/apis"
 import { createRaribleSdk } from "../index"
 import { createErc721V3Collection } from "../common/mint"
 import { MintResponseTypeEnum } from "../nft/mint"
+import { getEstimateGasInjects } from "../common/estimate-gas"
 import { cancel } from "./cancel"
 import { signOrder } from "./sign-order"
 import { UpsertOrder } from "./upsert-order"
@@ -40,7 +41,8 @@ describe.skip("cancel order", () => {
 
 	const getBaseOrderFee = async () => 0
 	const send = getSimpleSendWithInjects().bind(null, checkWalletChainId)
-	const orderService = new OrderFiller(ethereum, send, config, apis, getBaseOrderFee, env)
+	const estimateGas = getEstimateGasInjects()
+	const orderService = new OrderFiller(ethereum, send, estimateGas, config, apis, getBaseOrderFee, env)
 
 	const it = awaitAll({
 		testErc20: deployTestErc20(web3, "Test1", "TST1"),
@@ -107,7 +109,7 @@ describe.skip("cancel order", () => {
 		)
 
 		const order = await upserter.upsert({ order: form })
-		const tx = await cancel(checkLazyOrder, ethereum, send, config.exchange, checkWalletChainId, order)
+		const tx = await cancel(checkLazyOrder, ethereum, send, config.exchange, checkWalletChainId, apis, order)
 		await tx.wait()
 
 		const cancelledOrder = await retry(15, 2000, async () => {