+++ packages/sdk/src/order/fill-order/batch-purchase/batch-purchase.ts
@@ -16,6 +16,7 @@ import { checkChainId } from "../../check-chain-id"
 import type { IRaribleEthereumSdkConfig } from "../../../types"
 import type { EthereumNetwork } from "../../../types"
 import type {
+	AmmOrderFillRequest,
 	FillBatchOrderAction,
 	FillBatchOrderRequest,
 	FillBatchSingleOrderRequest,
@@ -30,8 +31,12 @@ import { OpenSeaOrderHandler } from "../open-sea"
 import { SeaportOrderHandler } from "../seaport"
 import { LooksrareOrderHandler } from "../looksrare"
 import { OriginFeeReducer } from "../common/origin-fee-reducer"
+import { X2Y2OrderHandler } from "../x2y2"
+import { AmmOrderHandler } from "../amm"
 import { createExchangeWrapperContract } from "../../contracts/exchange-wrapper"
 import type { SeaportV1OrderFillRequest } from "../types"
+import type { X2Y2OrderFillRequest } from "../types"
+import type { EstimateGasMethod } from "../../../common/estimate-gas"
 import { getUpdatedCalldata } from "../common/get-updated-call"
 import { getRequiredWallet } from "../../../common/get-required-wallet"
 
@@ -40,22 +45,27 @@ export class BatchOrderFiller {
 	openSeaHandler: OpenSeaOrderHandler
 	seaportHandler: SeaportOrderHandler
 	looksrareHandler: LooksrareOrderHandler
+	x2Y2Handler: X2Y2OrderHandler
+	ammHandler: AmmOrderHandler
 	private checkAssetType: CheckAssetTypeFunction
 	private checkLazyAssetType: (type: AssetType) => Promise<AssetType>
 
 	constructor(
 		private readonly ethereum: Maybe<Ethereum>,
 		private readonly send: SendFunction,
+		private readonly estimateGas: EstimateGasMethod,
 		private readonly config: EthereumConfig,
 		private readonly apis: RaribleEthereumApis,
 		private readonly getBaseOrderFee: (type: SimpleOrder["type"]) => Promise<number>,
 		private readonly env: EthereumNetwork,
 		private readonly sdkConfig?: IRaribleEthereumSdkConfig
 	) {
-		this.v2Handler = new RaribleV2OrderHandler(ethereum, send, config, getBaseOrderFee)
-		this.openSeaHandler = new OpenSeaOrderHandler(ethereum, send, config, apis, getBaseOrderFee, sdkConfig)
-		this.seaportHandler = new SeaportOrderHandler(ethereum, send, config, getBaseOrderFee, env)
-		this.looksrareHandler = new LooksrareOrderHandler(ethereum, send, config, getBaseOrderFee, env)
+		this.v2Handler = new RaribleV2OrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee)
+		this.openSeaHandler = new OpenSeaOrderHandler(ethereum, send, estimateGas, config, apis, getBaseOrderFee, sdkConfig)
+		this.seaportHandler = new SeaportOrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, env)
+		this.looksrareHandler = new LooksrareOrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, env)
+		this.x2Y2Handler = new X2Y2OrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, apis)
+		this.ammHandler = new AmmOrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, env)
 		this.checkAssetType = checkAssetType.bind(this, apis.nftCollection)
 		this.checkLazyAssetType = checkLazyAssetType.bind(this, apis.nftItem)
 		this.getTransactionRequestData = this.getTransactionRequestData.bind(this)
@@ -126,7 +136,9 @@ export class BatchOrderFiller {
 				request.order.type !== "RARIBLE_V2" &&
 				request.order.type !== "OPEN_SEA_V1" &&
 				request.order.type !== "LOOKSRARE" &&
-				request.order.type !== "SEAPORT_V1"
+				request.order.type !== "SEAPORT_V1" &&
+				request.order.type !== "X2Y2" &&
+				request.order.type !== "AMM"
 			) {
 				throw new Error("Unsupported order type for batch purchase")
 			}
@@ -201,7 +213,7 @@ export class BatchOrderFiller {
 			true // allowFail
 		)
 
-		await functionCall.estimateGas({
+		await this.estimateGas(functionCall, {
 			from: await ethereum.getFrom(),
 			value: totalValue.toString(),
 		})
@@ -240,6 +252,18 @@ export class BatchOrderFiller {
 			case "LOOKSRARE":
 				return this.looksrareHandler.getTransactionDataForExchangeWrapper(
 					<LooksrareOrderFillRequest>preparedOrder.request,
+					preparedOrder.request.originFees,
+					preparedOrder.fees
+				)
+			case "X2Y2":
+				return this.x2Y2Handler.getTransactionDataForExchangeWrapper(
+					<X2Y2OrderFillRequest>preparedOrder.request,
+					preparedOrder.request.originFees,
+					preparedOrder.fees
+				)
+			case "AMM":
+				return this.ammHandler.getTransactionDataForExchangeWrapper(
+					<AmmOrderFillRequest>preparedOrder.request,
 					preparedOrder.request.originFees,
 					preparedOrder.fees
 				)