+++ packages/sdk/src/order/fill-order/index.ts
@@ -9,6 +9,7 @@ import type {
 	SimpleOpenSeaV1Order,
 	SimpleOrder,
 	SimpleRaribleV2Order,
+	SimpleX2Y2Order,
 } from "../types"
 import type { SendFunction } from "../../common/send-transaction"
 import type { EthereumConfig } from "../../config/type"
@@ -19,6 +20,7 @@ import { checkLazyAssetType } from "../check-lazy-asset-type"
 import { checkChainId } from "../check-chain-id"
 import type { IRaribleEthereumSdkConfig } from "../../types"
 import type { EthereumNetwork } from "../../types"
+import type { EstimateGasMethod } from "../../common/estimate-gas"
 import type {
 	CryptoPunksOrderFillRequest,
 	FillOrderAction,
@@ -35,14 +37,19 @@ import type {
 	TransactionData,
 	SeaportV1OrderFillRequest,
 	SellOrderAction,
-	BuyOrderAction, LooksrareOrderFillRequest,
+	BuyOrderAction,
+	LooksrareOrderFillRequest,
+	X2Y2OrderFillRequest,
+	AmmOrderFillRequest,
 } from "./types"
 import { RaribleV1OrderHandler } from "./rarible-v1"
 import { RaribleV2OrderHandler } from "./rarible-v2"
 import { OpenSeaOrderHandler } from "./open-sea"
 import { CryptoPunksOrderHandler } from "./crypto-punks"
 import { SeaportOrderHandler } from "./seaport"
+import { X2Y2OrderHandler } from "./x2y2"
 import { LooksrareOrderHandler } from "./looksrare"
+import { AmmOrderHandler } from "./amm"
 
 export class OrderFiller {
 	v1Handler: RaribleV1OrderHandler
@@ -51,12 +58,15 @@ export class OrderFiller {
 	punkHandler: CryptoPunksOrderHandler
 	seaportHandler: SeaportOrderHandler
 	looksrareHandler: LooksrareOrderHandler
+	x2y2Handler: X2Y2OrderHandler
+	ammHandler: AmmOrderHandler
 	private checkAssetType: CheckAssetTypeFunction
 	private checkLazyAssetType: (type: AssetType) => Promise<AssetType>
 
 	constructor(
 		private readonly ethereum: Maybe<Ethereum>,
 		private readonly send: SendFunction,
+		private readonly estimateGas: EstimateGasMethod,
 		private readonly config: EthereumConfig,
 		private readonly apis: RaribleEthereumApis,
 		private readonly getBaseOrderFee: (type: SimpleOrder["type"]) => Promise<number>,
@@ -66,12 +76,38 @@ export class OrderFiller {
 		this.getBaseOrderFillFee = this.getBaseOrderFillFee.bind(this)
 		this.getTransactionData = this.getTransactionData.bind(this)
 		this.getBuyTx = this.getBuyTx.bind(this)
-		this.v1Handler = new RaribleV1OrderHandler(ethereum, apis.order, send, config, getBaseOrderFee, sdkConfig)
-		this.v2Handler = new RaribleV2OrderHandler(ethereum, send, config, getBaseOrderFee, sdkConfig)
-		this.openSeaHandler = new OpenSeaOrderHandler(ethereum, send, config, apis, getBaseOrderFee, sdkConfig)
-		this.punkHandler = new CryptoPunksOrderHandler(ethereum, send, config, getBaseOrderFee, sdkConfig)
-		this.seaportHandler = new SeaportOrderHandler(ethereum, send, config, getBaseOrderFee, env, sdkConfig)
-		this.looksrareHandler = new LooksrareOrderHandler(ethereum, send, config, getBaseOrderFee, env, sdkConfig)
+		this.v1Handler = new RaribleV1OrderHandler(
+			ethereum,
+			apis.order,
+			send,
+			estimateGas,
+			config,
+			getBaseOrderFee,
+			sdkConfig,
+		)
+		this.v2Handler = new RaribleV2OrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, sdkConfig)
+		this.openSeaHandler = new OpenSeaOrderHandler(ethereum, send, estimateGas, config, apis, getBaseOrderFee, sdkConfig)
+		this.punkHandler = new CryptoPunksOrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, sdkConfig)
+		this.seaportHandler = new SeaportOrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, env, sdkConfig)
+		this.looksrareHandler = new LooksrareOrderHandler(
+			ethereum,
+			send,
+			estimateGas,
+			config,
+			getBaseOrderFee,
+			env,
+			sdkConfig,
+		)
+		this.x2y2Handler = new X2Y2OrderHandler(ethereum, send, estimateGas, config, getBaseOrderFee, apis)
+		this.ammHandler = new AmmOrderHandler(
+			ethereum,
+			send,
+			estimateGas,
+			config,
+			getBaseOrderFee,
+			env,
+			sdkConfig
+		)
 		this.checkAssetType = checkAssetType.bind(this, apis.nftCollection)
 		this.checkLazyAssetType = checkLazyAssetType.bind(this, apis.nftItem)
 	}
@@ -84,7 +120,12 @@ export class OrderFiller {
 					if (!this.ethereum) {
 						throw new Error("Wallet undefined")
 					}
-					if (request.order.type === "SEAPORT_V1" || request.order.type === "LOOKSRARE") {
+					if (
+						request.order.type === "SEAPORT_V1" ||
+						request.order.type === "LOOKSRARE" ||
+						request.order.type === "X2Y2" ||
+						request.order.type === "AMM"
+					) {
 						return { request, inverted: request.order }
 					}
 					const from = toAddress(await this.ethereum.getFrom())
@@ -151,7 +192,11 @@ export class OrderFiller {
 			case "OPEN_SEA_V1":
 				return this.openSeaHandler.invert(<OpenSeaV1OrderFillRequest>request, from)
 			case "SEAPORT_V1":
-				throw new Error("Approve for Seaport orders is not implemented yet")
+				throw new Error("Invert for Seaport orders is not implemented yet")
+			case "X2Y2":
+				throw new Error("Invert for x2y2 orders is not implemented yet")
+			case "AMM":
+				throw new Error("Invert for AMM orders is not implemented yet")
 			case "CRYPTO_PUNK":
 				return this.punkHandler.invert(<CryptoPunksOrderFillRequest>request, from)
 			default:
@@ -169,6 +214,10 @@ export class OrderFiller {
 				return this.openSeaHandler.approve(inverted, isInfinite)
 			case "SEAPORT_V1":
 				throw new Error("Approve for Seaport orders is not implemented yet")
+			case "X2Y2":
+				throw new Error("Approve for x2y2 orders is not implemented yet")
+			case "AMM":
+				throw new Error("Approve for AMM orders is not implemented yet")
 			case "CRYPTO_PUNK":
 				return this.punkHandler.approve(inverted, isInfinite)
 			default:
@@ -198,8 +247,15 @@ export class OrderFiller {
 				)
 			case "LOOKSRARE":
 				return this.looksrareHandler.sendTransaction(<LooksrareOrderFillRequest>request)
+			case "X2Y2":
+				return this.x2y2Handler.fillOrder(
+					<SimpleX2Y2Order>request.order,
+					<X2Y2OrderFillRequest>request
+				)
 			case "CRYPTO_PUNK":
 				return this.punkHandler.sendTransaction(<SimpleCryptoPunkOrder>request.order, inverted)
+			case "AMM":
+				return this.ammHandler.sendTransaction(<AmmOrderFillRequest>request)
 			default:
 				throw new Error(`Unsupported order: ${JSON.stringify(inverted)}`)
 		}
@@ -230,6 +286,10 @@ export class OrderFiller {
 				return this.seaportHandler.getTransactionData(<SeaportV1OrderFillRequest>request)
 			case "LOOKSRARE":
 				return this.looksrareHandler.getTransactionData(<LooksrareOrderFillRequest>request)
+			case "AMM":
+				return this.ammHandler.getTransactionData(<AmmOrderFillRequest>request)
+			case "X2Y2":
+				return this.x2y2Handler.getTransactionData(<X2Y2OrderFillRequest>request)
 			case "CRYPTO_PUNK":
 				return this.punkHandler.getTransactionData(
           <SimpleCryptoPunkOrder>request.order,
@@ -274,6 +334,10 @@ export class OrderFiller {
 				return this.looksrareHandler.getOrderFee()
 			case "CRYPTO_PUNK":
 				return this.punkHandler.getOrderFee()
+			case "X2Y2":
+				return this.x2y2Handler.getOrderFee()
+			case "AMM":
+				return this.ammHandler.getOrderFee()
 			default:
 				throw new Error(`Unsupported order: ${JSON.stringify(order)}`)
 		}
@@ -293,6 +357,10 @@ export class OrderFiller {
 				return this.looksrareHandler.getBaseOrderFee()
 			case "CRYPTO_PUNK":
 				return this.punkHandler.getBaseOrderFee()
+			case "AMM":
+				return this.ammHandler.getBaseOrderFee()
+			case "X2Y2":
+				return this.ammHandler.getBaseOrderFee()
 			default:
 				throw new Error(`Unsupported order: ${JSON.stringify(order)}`)
 		}