+++ packages/sdk/src/order/fill-order/types.ts
@@ -11,6 +11,8 @@ import type {
 	SimpleOpenSeaV1Order,
 	SimpleRaribleV2Order,
 	SimpleSeaportV1Order,
+	SimpleX2Y2Order,
+	SimpleAmmOrder,
 } from "../types"
 import type { NftAssetType } from "../check-asset-type"
 
@@ -52,8 +54,11 @@ export type OpenSeaV1OrderFillRequest =
   Omit<CommonFillRequest<SimpleOpenSeaV1Order>, "amount"> & { payouts?: Part[], originFees?: Part[] }
 
 export type SeaportV1OrderFillRequest = CommonFillRequest<SimpleSeaportV1Order> & { originFees?: Part[] }
+export type X2Y2OrderFillRequest = CommonFillRequest<SimpleX2Y2Order> & { originFees?: Part[] }
 
 export type LooksrareOrderFillRequest = CommonFillRequest<SimpleLooksrareOrder> & { originFees?: Part[] }
+
+export type AmmOrderFillRequest = CommonFillRequest<SimpleAmmOrder> & { originFees?: Part[] }
 
 export type CryptoPunksOrderFillRequest = CommonFillRequest<SimpleCryptoPunkOrder>
 
@@ -64,8 +69,10 @@ export type SellOrderRequest =
 	RaribleV2OrderFillRequestV3Sell |
 	OpenSeaV1OrderFillRequest |
 	SeaportV1OrderFillRequest |
+	CryptoPunksOrderFillRequest |
 	LooksrareOrderFillRequest |
-	CryptoPunksOrderFillRequest
+	X2Y2OrderFillRequest |
+	AmmOrderFillRequest
 
 export type BuyOrderRequest =
 	LegacyOrderFillRequest |
@@ -73,8 +80,10 @@ export type BuyOrderRequest =
 	RaribleV2OrderFillRequestV3Buy |
 	OpenSeaV1OrderFillRequest |
 	SeaportV1OrderFillRequest |
+	CryptoPunksOrderFillRequest |
 	LooksrareOrderFillRequest |
-	CryptoPunksOrderFillRequest
+	X2Y2OrderFillRequest |
+	AmmOrderFillRequest
 
 export type FillOrderRequest =
   SellOrderRequest | BuyOrderRequest
@@ -85,7 +94,9 @@ export type FillBatchSingleOrderRequest =
 	// RaribleV2OrderFillRequestV3Buy |
 	OpenSeaV1OrderFillRequest |
 	LooksrareOrderFillRequest |
-	SeaportV1OrderFillRequest
+	SeaportV1OrderFillRequest |
+	X2Y2OrderFillRequest |
+	AmmOrderFillRequest
 
 export type FillBatchOrderRequest = FillBatchSingleOrderRequest[]
 
@@ -94,7 +105,8 @@ export enum ExchangeWrapperOrderType {
 	OPENSEA_V1 = 1,
 	SEAPORT_ADVANCED_ORDERS = 2,
 	X2Y2 = 3,
-	LOOKSRARE_ORDERS = 4
+	LOOKSRARE_ORDERS = 4,
+	AAM = 5
 }
 
 export type PreparedOrderRequestDataForExchangeWrapper = {