+++ packages/test-common/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.92](https://github.com/rarible/ethereum-sdk/compare/v0.12.91...v0.12.92) (2022-09-13)
+
+**Note:** Version bump only for package @rarible/ethereum-sdk-test-common
+
+
+
+
+
+## [0.12.89](https://github.com/rarible/ethereum-sdk/compare/v0.12.88...v0.12.89) (2022-09-06)
+
+**Note:** Version bump only for package @rarible/ethereum-sdk-test-common
+
+
+
+
+
 ## [0.12.87](https://github.com/rarible/ethereum-sdk/compare/v0.12.86...v0.12.87) (2022-08-22)
 
 **Note:** Version bump only for package @rarible/ethereum-sdk-test-common